backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.41
aiosqlite==0.21.0
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
anthropic==0.52.0
python-multipart==0.0.20
fastembed
httpx
PyPDF2
python-pptx
youtube-transcript-api
